api/requirements.txt
fastapi==0.115.7
uvicorn==0.34.0
requests
python-multipart
transformers==4.48.1
torch==2.6.0
datasets==3.2.0
python-dotenv==1.0.1
PyPDF2==3.0.1
huggingface_hub==0.28.1
hypercorn==0.13.1 
pdfplumber==0.11.5

# YOLO / Computer Vision
ultralytics>=8.0.0
opencv-python-headless>=4.8.0
Pillow>=10.0.0

# Database (Fase 2)
# sqlalchemy>=2.0.0
# asyncpg>=0.29.0
# alembic>=1.13.0

# Scraping (Fase 2)
# beautifulsoup4>=4.12.0
# playwright>=1.40.0
# httpx>=0.26.0

# Cache (Fase 2)
# redis>=5.0.0

# TEST
pytest==7.1.2
pytest-asyncio>=0.23.0